add normal, uvs (basic) and color (import only) handling to maya plugin
